Edit cell
A widget can be edited. An unpublished widget will have its contents
changed, while a published widget will remain the same with a new widget
added to the dashboard with the new changes.
In this iteration a user must again edit a "forked" widget to publish
it, since that option is not available when editing a published cell.